+++ src/Module/Colony/Action/DeactivateBuilding/DeactivateBuilding.php
@@ -42,7 +42,7 @@ final class DeactivateBuilding implements ActionControllerInterface
 
         $field = $this->planetFieldRepository->getByColonyAndFieldId(
             $colony->getId(),
-            (int) request::indInt('fid')
+            request::indInt('fid')
         );
 
         if ($field === null) {
