+++ src/Module/Ship/Action/Colonize/Colonize.php
@@ -94,8 +94,8 @@ final class Colonize implements ActionControllerInterface
             $userId
         );
 
-        $colonyId = (int)request::getIntFatal('colid');
-        $fieldId = (int)request::getIntFatal('field');
+        $colonyId = request::getIntFatal('colid');
+        $fieldId = request::getIntFatal('field');
 
         $colony = $this->colonyRepository->find($colonyId);
         $field = $this->planetFieldRepository->find($fieldId);
