+++ src/Module/Ship/View/ShowBeamFromColony/ShowBeamFromColony.php
@@ -44,7 +44,7 @@ final class ShowBeamFromColony implements ViewControllerInterface
         $game->setPageTitle('Von Kolonie beamen');
         $game->setMacroInAjaxWindow('html/shipmacros.xhtml/entity_not_available');
 
-        $target = $this->colonyRepository->find((int)request::getIntFatal('target'));
+        $target = $this->colonyRepository->find(request::getIntFatal('target'));
         if ($target === null || !InteractionChecker::canInteractWith($ship, $target, $game, true)) {
             return;
         }
