+++ src/Module/Station/Action/DeleteDockPrivilege/DeleteDockPrivilege.php
@@ -37,7 +37,7 @@ final class DeleteDockPrivilege implements ActionControllerInterface
         );
 
         $game->setView(ShowDockingPrivileges::VIEW_IDENTIFIER);
-        $privilege = $this->dockingPrivilegeRepository->find((int) request::getIntFatal('privilegeid'));
+        $privilege = $this->dockingPrivilegeRepository->find(request::getIntFatal('privilegeid'));
 
         if ($privilege->getShip()->getId() !== $ship->getId()) {
             return;
